tokens

+++ src/main/java/com/example/reversed/services/TaskServiceImpl.java
@@ -2,9 +2,14 @@ package com.example.reversed.services;
 
 import com.example.reversed.beans.Group;
 import com.example.reversed.beans.Task;
+import com.example.reversed.beans.User;
+import com.example.reversed.beans.dtos.TaskDTO;
 import com.example.reversed.exceptions.ExceptionState;
 import com.example.reversed.exceptions.TaskException;
+import com.example.reversed.mappers.Mapper;
 import com.example.reversed.repos.TaskRepository;
+import com.example.reversed.repos.UserRepository;
+import com.example.reversed.security.TokenManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,30 +23,43 @@ import static com.example.reversed.exceptions.ExceptionState.NOT_EXISTS;
 public class TaskServiceImpl implements TaskService{
 
     private final TaskRepository taskRepository;
+    private final Mapper mapper;
+    private final TokenManager tokenManager;
+    private final UserRepository userRepository;
 
     @Override
-    public void add(Task task) throws TaskException {
+    public TaskDTO add(TaskDTO task, int userID) throws TaskException {
 
         if (taskRepository.existsById(task.getId())){
             throw new TaskException(ExceptionState.ALREADY_EXISTS);
         }
 
-        taskRepository.save(task);
+        User user = userRepository.findById(userID).orElseThrow(() -> new TaskException(ExceptionState.NOT_EXISTS));
+
+        Task t1 = mapper.mapToEntity(task);
+        t1.setUser(user);
+
+        return mapper.mapToObject(taskRepository.saveAndFlush(t1));
     }
 
     @Override
-    public void update(long id, Task task) throws TaskException {
+    public TaskDTO update(long id, TaskDTO task, int userID) throws TaskException {
 
         if (!taskRepository.existsById(id)){
             throw new TaskException(NOT_EXISTS);
         }
 
-        taskRepository.saveAndFlush(task);
+        User user = userRepository.findById(userID).orElseThrow(() -> new TaskException(ExceptionState.NOT_EXISTS));
 
+        Task t1 = mapper.mapToEntity(task);
+        t1.setId(id);
+        t1.setUser(user);
+
+        return mapper.mapToObject(taskRepository.saveAndFlush(t1));
     }
 
     @Override
-    public void delete(long id) throws TaskException {
+    public void delete(long id, int userID) throws TaskException {
 
         //TODO - this is the same as update, so need to export it to util
 
@@ -49,38 +67,41 @@ public class TaskServiceImpl implements TaskService{
             throw new TaskException(NOT_EXISTS);
         }
 
+        if (!userRepository.existsById(userID)) {
+            throw new TaskException(ExceptionState.NOT_EXISTS);
+        }
+
         taskRepository.deleteById(id);
-
     }
 
     @Override
-    public Task getOne(long id) throws TaskException {
-        return taskRepository.findById(id).orElseThrow(() -> new TaskException(NOT_EXISTS));
+    public TaskDTO getOne(long id) throws TaskException {
+        return mapper.mapToObject(taskRepository.findById(id).orElseThrow(() -> new TaskException(NOT_EXISTS)));
     }
 
     @Override
-    public List<Task> getAll() {
-        return taskRepository.findAll();
+    public List<TaskDTO> getAll(int userID) {
+        return mapper.mapToObjectList(taskRepository.findByUserId(userID));
     }
 
     @Override
-    public Task getByTitle(String title) throws TaskException {
-        return taskRepository.findByTitle(title).orElseThrow(() -> new TaskException(NOT_EXISTS));
+    public TaskDTO getByTitle(String title) throws TaskException {
+        return mapper.mapToObject(taskRepository.findByTitle(title).orElseThrow(() -> new TaskException(NOT_EXISTS)));
     }
 
     @Override
-    public List<Task> getByGroup(Group group) {
-        return taskRepository.findByGroupType(group);
+    public List<TaskDTO> getByGroup(Group group) {
+        return mapper.mapToObjectList(taskRepository.findByGroupType(group));
     }
 
     @Override
-    public List<Task> getBefore(Date date) {
-        return taskRepository.findByWhenToDoBefore(date);
+    public List<TaskDTO> getBefore(Date date) {
+        return mapper.mapToObjectList(taskRepository.findByWhenToDoBefore(date));
     }
 
     @Override
-    public List<Task> getAfter(Date date) {
-        return taskRepository.findByWhenToDoAfter(date);
+    public List<TaskDTO> getAfter(Date date) {
+        return mapper.mapToObjectList(taskRepository.findByWhenToDoAfter(date));
     }
 
     @Override
@@ -89,17 +110,17 @@ public class TaskServiceImpl implements TaskService{
     }
 
     @Override
-    public List<Task> sortByNearest() {
-        return taskRepository.sortByNearest();
+    public List<TaskDTO> sortByNearest() {
+        return mapper.mapToObjectList(taskRepository.sortByNearest());
     }
 
     @Override
-    public List<Task> sortByFarthest() {
-        return taskRepository.sortByFarthest();
+    public List<TaskDTO> sortByFarthest() {
+        return mapper.mapToObjectList(taskRepository.sortByFarthest());
     }
 
     @Override
-    public List<Task> getAllExpired() {
-        return taskRepository.getExpired();
+    public List<TaskDTO> getAllExpired() {
+        return mapper.mapToObjectList(taskRepository.getExpired());
     }
 }
